+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 1.4.3
+
+  * Added integration tests for Unicode support
+  * Added use of sosueme/conf/dot-factual; demo and tests now expect auth in ~/.factual/factual-auth.yaml
+
 ## 1.4.2
 
   * Added support for Match